+++ src/variables/constants.js
@@ -57,7 +57,7 @@ const TIMEBOMB_ACTIVATED = true;
  *
  * To see the date in **90** days, [click here](https://www.google.com/search?q=Date+90+days+from+today).
  */
-const EXPIRY_DATE = new Date(2025, 6, 1);
+const EXPIRY_DATE = new Date(2025, 2, 1);
 
 module.exports = {
   GLOBAL_NAME,